Re-allow pick Seraphim in lobby
There is no need to restrict Seraphim in the lobby, with upcoming
Seraphim mission it's welcome change. Reason that in FA you could not
play seraphim is poor, that way we would not allow any faction since in
vanilla campaign you could play just one faction in each mission.
So trying to pick Seraphim into FA campaign is same as picking UEF into
Cybran mission.

In future we could maybe store allowed factions with the map and change
the lobby according to that.