Add support any number of players
Fixes #36 and  #37
It requires players' armies to be called `Player1, Player2, ... PlayerN`
Number of slots depends on the mission, no longer fixed to 4.

This can be added once the current missions are changed to work with
this, which means renaming `Coop` armies to `Player`

Later the campaign code will needs to be changed since it's using army
`Player` on many places, which should get renamed to `Player1` to keep
the naming of the armies sane.