Improve BrainsCompareNumCategory
I added the possibility to use HumanPlayers together with other armies and it doesn't need to be in the first position of the list anymore. I made sure to remove all duplicates aswell so you don't count units of armies multiple times. (for example: when you make an error like this -> TargetBrains={HumanPlayers, Player, Cybran}